+++ eval_protocol/pytest/__init__.py
@@ -1,18 +1,17 @@
-from .default_agent_rollout_processor import default_agent_rollout_processor
+from .default_agent_rollout_processor import AgentRolloutProcessor
+from .default_base_rollout_process import BaseRolloutProcessor
 from .default_dataset_adapter import default_dataset_adapter
-from .default_mcp_gym_rollout_processor import default_mcp_gym_rollout_processor
-from .default_no_op_rollout_process import default_no_op_rollout_processor
-from .default_single_turn_rollout_process import default_single_turn_rollout_processor
+from .default_mcp_gym_rollout_processor import MCPGymRolloutProcessor
+from .default_single_turn_rollout_process import SingleTurnRolloutProcessor
 from .evaluation_test import evaluation_test
-from .types import RolloutProcessor, RolloutProcessorConfig
+from .types import RolloutProcessorConfig
 
 __all__ = [
-    "default_agent_rollout_processor",
-    "default_mcp_gym_rollout_processor",
-    "default_no_op_rollout_processor",
-    "default_single_turn_rollout_processor",
+    "AgentRolloutProcessor",
+    "MCPGymRolloutProcessor",
+    "BaseRolloutProcessor",
+    "SingleTurnRolloutProcessor",
     "default_dataset_adapter",
-    "RolloutProcessor",
     "RolloutProcessorConfig",
     "evaluation_test",
 ]
