+++ tests/pytest/test_lunar_lander.py
@@ -9,7 +9,7 @@ from typing import Any, Dict, List
 
 from eval_protocol.models import EvaluateResult, EvaluationRow, InputMetadata, Message
 from eval_protocol.pytest import evaluation_test
-from eval_protocol.pytest.default_mcp_gym_rollout_processor import default_mcp_gym_rollout_processor
+from eval_protocol.pytest.default_mcp_gym_rollout_processor import MCPGymRolloutProcessor
 
 
 def lunar_lander_to_evaluation_row(data: List[Dict[str, Any]]) -> List[EvaluationRow]:
@@ -39,7 +39,7 @@ def lunar_lander_to_evaluation_row(data: List[Dict[str, Any]]) -> List[Evaluatio
     input_dataset=["tests/pytest/data/lunar_lander_dataset.jsonl"],
     dataset_adapter=lunar_lander_to_evaluation_row,
     completion_params=[{"temperature": 0.0, "max_tokens": 4096, "model": "gpt-4.1"}],
-    rollout_processor=default_mcp_gym_rollout_processor,
+    rollout_processor=MCPGymRolloutProcessor(),
     passed_threshold=0.0,
     num_runs=1,
     mode="pointwise",
