+++ tests/pytest/test_pytest_ids.py
@@ -3,7 +3,7 @@ from typing import List
 import eval_protocol.dataset_logger as dataset_logger
 from eval_protocol.dataset_logger.dataset_logger import DatasetLogger
 from eval_protocol.models import EvaluationRow
-from eval_protocol.pytest.default_no_op_rollout_process import default_no_op_rollout_processor
+from eval_protocol.pytest.default_base_rollout_process import BaseRolloutProcessor
 from tests.pytest.test_markdown_highlighting import markdown_dataset_to_evaluation_row
 
 
@@ -30,7 +30,7 @@ async def test_evaluation_test_decorator(monkeypatch):
         ],
         completion_params=[{"temperature": 0.0, "model": "dummy/local-model"}],
         dataset_adapter=markdown_dataset_to_evaluation_row,
-        rollout_processor=default_no_op_rollout_processor,
+        rollout_processor=BaseRolloutProcessor(),
         mode="pointwise",
         combine_datasets=False,
         num_runs=2,
@@ -71,7 +71,7 @@ async def test_evaluation_test_decorator_ids_single(monkeypatch):
             {"temperature": 1.0, "model": "dummy/local-model"},
         ],
         dataset_adapter=markdown_dataset_to_evaluation_row,
-        rollout_processor=default_no_op_rollout_processor,
+        rollout_processor=BaseRolloutProcessor(),
         mode="pointwise",
         combine_datasets=False,
         num_runs=5,
