+++ tests/pytest/test_pytest_mcp_config.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from typing import List
 
 from eval_protocol.models import EvaluateResult, EvaluationRow, Message
-from eval_protocol.pytest import default_agent_rollout_processor, evaluation_test
+from eval_protocol.pytest import AgentRolloutProcessor, evaluation_test
 
 
 @evaluation_test(
@@ -19,7 +19,7 @@ from eval_protocol.pytest import default_agent_rollout_processor, evaluation_tes
             )
         ]
     ],
-    rollout_processor=default_agent_rollout_processor,
+    rollout_processor=AgentRolloutProcessor(),
     completion_params=[{"model": "fireworks_ai/accounts/fireworks/models/gpt-oss-20b"}],
     mode="pointwise",
     mcp_config_path="tests/pytest/mcp_configurations/mock_discord_mcp_config.json",
